+++ lib/article_json.rb
@@ -46,6 +46,7 @@ require_relative 'article_json/export/common/html/elements/text'
 require_relative 'article_json/export/common/html/elements/heading'
 require_relative 'article_json/export/common/html/elements/paragraph'
 require_relative 'article_json/export/common/html/elements/list'
+require_relative 'article_json/export/common/html/elements/quote'
 
 require_relative 'article_json/export/html/elements/base'
 require_relative 'article_json/export/html/elements/text'
